requirements.txt
qrcode==7.4.2
Pillow==10.0.0
tensorflow==2.12.0
tensorflow-model-optimization==0.8.0
numpy>=1.20.0
scipy>=1.7.0
matplotlib>=3.5.0 